Fix vertex degree filtration (#26)
* Update flagser to latest release

This release fix the bug encountered with vertex_degree filtration

* Update filtration tests with vertex_degree

* Refactor conftest.py for better parametrizations

The new fixture `filtration` is introduced alongside `flag_file` for all tests. If it is present then `flag_file` has to be in `files_with_filtration_results` which is currently set to`["d5.flag"]`. This way we avoid skipping tests but can parametrize wrt the filtration in the case of d5.flag.

Signed-off-by: julian <julian.burellaperez@heig-vd.ch>
Co-authored-by: Umberto <u.lupo@l2f.ch>